pyproject.toml
[project]
name = "liquidation-alerter"
version = "0.1.0"
description = "DeFi Liquidation Alerter - Monitor positions and receive Telegram alerts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "web3>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "aiohttp>=3.9.0",
    "python-telegram-bot>=20.0",
    "prometheus-client>=0.19.0",
    "greenlet>=3.0.0",
]

[project.scripts]
liquidation-alerter = "app.main:main"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning:websockets.legacy",
]
